NX-364: Remove enunciate
Removed all the DocumentationBundles from OSS,
and removed deprecated AbstractDocumentationResourceBundle
while the DocumentationBundle interface and
DocumentationBundleSupport left in place.

Also, related IT modified to NOT check for documentation
while we do not have one.
