+++ components/nexus-plugin-api/src/main/java/org/sonatype/nexus/plugin/support/DocumentationBundleSupport.java
@@ -10,9 +10,20 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
+
 package org.sonatype.nexus.plugin.support;
 
+import java.net.URL;
+import java.util.Enumeration;
+import java.util.LinkedList;
+import java.util.List;
+
+import org.sonatype.nexus.mime.MimeSupport;
 import org.sonatype.nexus.plugin.PluginIdentity;
+import org.sonatype.nexus.web.WebResource;
+import org.sonatype.sisu.goodies.common.ComponentSupport;
+
+import org.eclipse.sisu.space.ClassSpace;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -22,12 +33,56 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * @since 2.7
  */
 public abstract class DocumentationBundleSupport
-    extends AbstractDocumentationResourceBundle
+    extends ComponentSupport
+    implements DocumentationBundle
 {
+  private final MimeSupport mimeSupport;
+
+  private final ClassSpace space;
+
   private final PluginIdentity owner;
 
-  protected DocumentationBundleSupport(final PluginIdentity plugin) {
-    this.owner = checkNotNull(plugin);
+  protected DocumentationBundleSupport(final MimeSupport mimeSupport, final ClassSpace space,
+                                       final PluginIdentity owner)
+  {
+    this.mimeSupport = checkNotNull(mimeSupport);
+    this.space = checkNotNull(space);
+    this.owner = checkNotNull(owner);
+  }
+
+  @Override
+  public List<WebResource> getResources() {
+    List<WebResource> resources = new LinkedList<WebResource>();
+
+    if (space != null) {
+      for (Enumeration<URL> e = space.findEntries("docs", "*", true); e.hasMoreElements(); ) {
+        URL url = e.nextElement();
+
+        String name = url.getPath();
+        name = name.substring(1 + name.lastIndexOf("/docs/"), name.length());
+
+        // to lessen clash possibilities, this way only within single plugin may be clashes, but one
+        // plugin is usually developed by one team or one user so this is okay
+        // system-wide clashes are much harder to resolve
+        String path = "/" + getPluginId() + "/" + getPathPrefix() + "/" + name;
+
+        resources.add(new UrlWebResource(url, path, mimeSupport.guessMimeTypeFromPath(name)));
+      }
+    }
+
+    if (log.isTraceEnabled()) {
+      log.trace("Discovered documentation for: {}", getPluginId());
+      for (WebResource resource : resources) {
+        log.trace("  {}", resource);
+      }
+    }
+
+    return resources;
+  }
+
+  @Override
+  public String getPathPrefix() {
+    return "default";
   }
 
   @Override
