--- plugins/basic/nexus-timeline-plugin/src/main/java/org/sonatype/nexus/timeline/internal/TimelineDocumentationResourceBundle.java
@@ -1,36 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2007-2014 Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-
-package org.sonatype.nexus.timeline.internal;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import org.sonatype.nexus.plugin.support.DocumentationBundleSupport;
-import org.sonatype.nexus.timeline.TimelinePlugin;
-
-/**
- * Documentation bundle for timeline.
- */
-// TODO: Enunciate craps out on timeline due to a bug (two xml-ish @produces on same resource)
-//@Named
-//@Singleton
-public class TimelineDocumentationResourceBundle
-    extends DocumentationBundleSupport
-{
-  @Inject
-  public TimelineDocumentationResourceBundle(final TimelinePlugin plugin) {
-    super(plugin);
-  }
-}
