--- plugins/basic/nexus-unpack-plugin/src/main/java/org/sonatype/nexus/unpack/internal/UnpackDocumentationResourceBundle.java
@@ -1,34 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2007-2014 Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package org.sonatype.nexus.unpack.internal;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import org.sonatype.nexus.plugin.support.DocumentationBundleSupport;
-import org.sonatype.nexus.unpack.UnpackPlugin;
-
-/**
- * Unpack plugin documentation bundle.
- */
-@Named
-@Singleton
-public class UnpackDocumentationResourceBundle
-    extends DocumentationBundleSupport
-{
-  @Inject
-  public UnpackDocumentationResourceBundle(final UnpackPlugin plugin) {
-    super(plugin);
-  }
-}
