+++ testsuite/legacy-testsuite/src/test/java/org/sonatype/nexus/testsuite/plugin/nxcm2124/NXCM2124CheckConsoleDocumentationIT.java
@@ -49,6 +49,10 @@ public class NXCM2124CheckConsoleDocumentationIT
 
     assertThat(getPluginsNames(pluginInfos), hasItem(pluginName));
 
+    // TODO: Since NX-364 Enunciate documentation is removed, hence the
+    // part of this IT that checks for it is also disabled until we provide
+    // some alternative solution for doco.
+    /*
     PluginInfoDTO pluginConsolePlugin =
         this.getPluginInfoByName(pluginInfos, pluginName);
     Assert.assertNotNull(pluginConsolePlugin.getDocumentation());
@@ -65,5 +69,6 @@ public class NXCM2124CheckConsoleDocumentationIT
     finally {
       RequestFacade.releaseResponse(r);
     }
+    */
   }
 }
